backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
requests==2.31.0
Pillow==10.4.0
python-dotenv==1.0.1
